+++ changelog.md
@@ -2,6 +2,7 @@
 
 ## v0.2.0 | In development
 
+* Channel slices can be downsampled: `lf_force = hf_force.downsampled_by(20)`.
 * `h5py` >= v2.8 is now required.
 
 ## v0.1.0 | 2018-06-20